requirements.txt
jira==3.4.1
click~=8.1.3
Jinja2~=3.1.2
pyfiglet~=0.8.post1
PyYAML~=6.0
jsonschema~=4.16.0
inquirerpy~=0.3.4
termcolor==2.0.1
python-editor~=1.0.4
